requirements.txt
asgiref==3.11.0
attrs==25.4.0
autobahn==25.11.1
Automat==25.4.16
cbor2==5.7.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
defusedxml==0.7.1
Django==5.2.9
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
django-cors-headers==4.3.1
django-storages==1.14.2
boto3==1.34.34
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
msgpack==1.1.2
oauthlib==3.3.1
packaging==25.0
psycopg2==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
service-identity==24.2.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.4
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.5.0
zope.interface==8.1.1
